arreglando componentes
