--- burguer-queen/src/components/header.jsx
@@ -1,23 +1,0 @@
-import React from "react";
-import image from "../images/Gourmet.jpg";
-import imageIcon from "../images/cerrar-sesion.png";
-import { useNavigate } from "react-router-dom";
-
-function Header(){
-    const navigate = useNavigate();
-
-    const buttonOut = (e) => {
-      e.preventDefault();
-      navigate("/");
-    };
-  
-    return (
-      <header id="containerHeader">
-        <img className="logoImageAdmin" src={image} alt="foto del logo" />
-        <h1 className="titleAdmin"> Gourmet Queen </h1>
-        <img id="logoutIcon" onClick={buttonOut} src={imageIcon} alt="logout" />
-      </header>
-    );
-  }
-  
-  export { Header };
